+++ src/pwd/pwd.rs
@@ -1,5 +1,5 @@
 #![crate_name = "pwd"]
-#![feature(collections, core, io, os, rustc_private)]
+#![feature(collections, core, old_io, os, rustc_private)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -22,7 +22,7 @@ mod util;
 static NAME: &'static str = "pwd";
 static VERSION: &'static str = "1.0.0";
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let program = args[0].clone();
     let opts = [
         getopts::optflag("", "help", "display this help and exit"),
