+++ src/relpath/relpath.rs
@@ -1,5 +1,5 @@
 #![crate_name = "relpath"]
-#![feature(collections, core, os, path, rustc_private)]
+#![feature(collections, core, os, old_path, rustc_private)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -20,7 +20,7 @@ use getopts::{optflag, optopt, getopts, usage};
 static NAME: &'static str = "relpath";
 static VERSION: &'static str = "1.0.0";
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let program = &args[0];
     let options = [
         optflag("h", "help", "Show help and exit"),
